Removed deprecated code.
