+++ .travis.yml
@@ -3,8 +3,5 @@ sudo: false
 node_js:
   - "0.12"
   - "0.10"
-  - "iojs-v1"
-  - "iojs-v2"
-  - "iojs-v3"
   - "4.0"
   - "4.1"
