+++ test/mqtt.js
@@ -6,8 +6,6 @@
  */
 
 var fs = require('fs'),
-  net = require('net'),
-  sinon = require('sinon'),
   mqtt = require('../');
 
 /**
@@ -53,7 +51,6 @@ describe('mqtt', function () {
     });
 
     it('should return an MqttClient with correct host when called with a host and port', function () {
-      sinon.spy(mqtt, 'createClient');
       var c = mqtt.connect('tcp://user:pass@localhost:1883');
 
       c.options.should.have.property('hostname', 'localhost');
@@ -165,98 +162,6 @@ describe('mqtt', function () {
       c.on('error', function () {});
 
       c.should.be.instanceOf(mqtt.MqttClient);
-    });
-  });
-
-  describe('#createClient', function () {
-    it('should return an MqttClient', function () {
-      var c = mqtt.createClient();
-
-      c.should.be.instanceOf(mqtt.MqttClient);
-    });
-  });
-
-  describe('#createSecureClient', function () {
-    it('should return an MqttClient', function () {
-      var c = mqtt.createSecureClient();
-
-      c.on('error', function () {});
-
-      c.should.be.instanceOf(mqtt.MqttClient);
-    });
-
-    it('should support passing the key and cert', function () {
-      var c = mqtt.createSecureClient({
-        keyPath: __dirname + '/helpers/private-key.pem',
-        certPath: __dirname + '/helpers/public-cert.pem'
-      });
-
-      c.on('error', function () {});
-
-      c.should.be.instanceOf(mqtt.MqttClient);
-    });
-
-    it('should throw on incorrect args');
-  });
-
-  describe('#createSecureClientWithListCA', function () {
-    it('should return an MqttClient', function () {
-      var c = mqtt.createSecureClient();
-
-      c.on('error', function () {});
-
-      c.should.be.instanceOf(mqtt.MqttClient);
-    });
-
-    it('should support passing the key, cert and CA list', function () {
-      var c = mqtt.createSecureClient({
-        keyPath: __dirname + '/helpers/private-key.pem',
-        certPath: __dirname + '/helpers/public-cert.pem',
-        ca: [__dirname + '/helpers/public-cert.pem']
-      });
-
-      c.on('error', function () {});
-
-      c.should.be.instanceOf(mqtt.MqttClient);
-    });
-
-    it('should throw on incorrect args');
-  });
-
-  describe('#createConnection', function () {
-    before(function () {
-      // Setup dummy server
-
-      // If there's an error it's probably EADDRINUSE
-      // Just use whatever's there already (likely mosquitto)
-      this.server = new net.Server();
-      this.server.listen(1883);
-      this.server.on('error', function () {});
-    });
-    it('should return an MqttConnection', function () {
-      var c = mqtt.createConnection();
-
-      c.should.be.instanceOf(mqtt.MqttConnection);
-    });
-
-    it('should fire callback on net connect', function (done) {
-      mqtt.createConnection(done);
-    });
-    it('should bind stream close to connection', function (done) {
-      var c = mqtt.createConnection();
-      c.once('connected', function () {
-        c.once('close', function () {
-          done();
-        });
-        c.stream.end();
-      });
-    });
-    it('should bind stream error to conn', function (done) {
-      var c = mqtt.createConnection();
-      c.once('error', function () {
-        done();
-      });
-      c.stream.emit('error', new Error('Bad idea!'));
     });
   });
 });
